refactor: move webnode_parser to nlp solution
